Fix outputJSON discrepancies
