+++ website/src/components/Visualisations/IITO/Util/ForceGraphGenerator.tsx
@@ -2,7 +2,7 @@
 // TypeScript does not work nicely at all with d3 so need to come back and fix these
 import * as d3 from 'd3'
 import styles from '../IITO.module.css'
-import { Link, Node } from '../../../../consts/types'
+import { Link, Node } from './ForceGraph'
 
 export default function runForceGraph(
   container: any,
